Correction JSON (les sous object ne doivent pas être entre guillemets) Utilisation d'un parser JSON plutot que de le faire a la main dans getStructureId et getStructureName

+++ Android/src/com/google/code/sig_1337/model/ISig1337.java
@@ -58,7 +58,7 @@ public interface ISig1337 {
 	 *            latitude.
 	 * @return the id of the structure.
 	 */
-	public int getStructureId(double x, double y);
+	public long getStructureId(double x, double y);
 
 	/**
 	 * Get the name of the structure at the given coordinates.
